Keep logged in if the device is rotated
Rotations were causing recompositions, thus resetting the default value for `isLoggedIn` in the composable.

I created a view model to hold manage `isLoggedIn`.
The view model will be instantiated in a component at some point, thus it needs to survive recompositions, otherwise the view model will always be using the default values, effectively.
Therefore, simply moving the state to it won't work for the same reason as the original problem.
This is why I used `viewModel<_>()` to create it.

The view model will survive recompositions and its state can be shared with and observed from multiple composables.